프로필 첨부파일 페이지 이동 시 사진 삭제되는 것 코드 수정

+++ dang/src/main/webapp/WEB-INF/views/template/dang_side_profile.jsp
@@ -838,7 +838,7 @@
 		function deleteAttachmentNo(){
 			var newAttachmentNo = $("[name=attachmentNo]").val();
 			var originAttachmentNo = $("[name=originAttachmentNo]").val();
-			if(newAttachmentNo!=originAttachmentNo){ //새로 사진등록한 상태
+			if(newAttachmentNo!=originAttachmentNo && originAttachmentNo!=""){ //새로 사진등록한 상태
 				$.ajax({
 					url:"${pageContext.request.contextPath}/rest_attachment/delete/"+newAttachmentNo,
 					method:"delete",
